requirements.txt
django-anymail
